+++ src/util.spec.ts
@@ -17,6 +17,20 @@ beforeEach(() => {
 
 describe('Util', () => {
 
+    describe('isAssignableExpression', () => {
+        it('works', () => {
+            expect(util.isAssignableExpression('function test(): endFunction')).to.be.false;
+            expect(util.isAssignableExpression('if true then print true')).to.be.false;
+            expect(util.isAssignableExpression('while true: print true')).to.be.false;
+            expect(util.isAssignableExpression('a.b.c = "test"')).to.be.false;
+            expect(util.isAssignableExpression('array[0] = "test"')).to.be.false;
+            expect(util.isAssignableExpression('2+2')).to.be.true;
+            expect(util.isAssignableExpression('createThing()')).to.be.true;
+            expect(util.isAssignableExpression('a.b.c')).to.be.true;
+            expect(util.isAssignableExpression('array[0]')).to.be.true;
+        });
+    });
+
     describe('removeTrailingNewline', () => {
         it('works', () => {
             expect(util.removeTrailingNewline('\r\n')).to.equal('');